refactor: PROD-819 분석 경계 단순화
