+++ apps/app/src/analytics/route.test.ts
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
-import { isNewRouteTemplate, normalizeRouteTemplate } from './route';
+import { normalizeRouteTemplate } from './route';
 
 describe('normalizeRouteTemplate', () => {
   it('route group과 index를 제거하고 dynamic segment placeholder를 보존한다', () => {
@@ -22,11 +22,5 @@ describe('normalizeRouteTemplate', () => {
   it('empty segment는 root route로 정규화한다', () => {
     assert.equal(normalizeRouteTemplate([]), '/');
     assert.equal(normalizeRouteTemplate(['(tabs)', 'index']), '/');
-  });
-
-  it('same template re-render는 dedupe하고 template 전환은 capture한다', () => {
-    assert.equal(isNewRouteTemplate(null, '/'), true);
-    assert.equal(isNewRouteTemplate('/', '/'), false);
-    assert.equal(isNewRouteTemplate('/', '/privacy'), true);
   });
 });
