+++ apps/app/src/components/AppProviders.tsx
@@ -1,7 +1,6 @@
-import { Suspense, useEffect } from 'react';
+import { Suspense } from 'react';
 import { AnalyticsPageviewBridge } from '@/analytics/AnalyticsPageviewBridge';
 import { AnalyticsSessionBridge } from '@/analytics/AnalyticsSessionBridge';
-import { initializeAnalytics } from '@/analytics/client';
 import { RelayActorProvider, useRelayActor } from '@/relay/RelayActorProvider';
 import {
   SessionErrorProvider,
@@ -41,10 +40,6 @@ function RelaySessionBoundary({ children }: PropsWithChildren) {
 }
 
 export function AppProviders({ children }: PropsWithChildren) {
-  useEffect(() => {
-    initializeAnalytics();
-  }, []);
-
   return (
     <ThemeProvider mode="light">
       <ToastProvider>
